+++ controller/scheduler.c
@@ -89,7 +89,7 @@ void rescheduleTask(cfTaskId_e taskId, uint32_t newPeriodMicros)
         taskId = currentTaskId;
 
     if (taskId < TASK_COUNT) {
-        cfTasks[taskId].desiredPeriod = MAX(100, newPeriodMicros);  // Limit delay to 100us (10 kHz) to prevent scheduler clogging
+        cfTasks[taskId].desiredPeriod = MAX((uint32_t)100, newPeriodMicros);  // Limit delay to 100us (10 kHz) to prevent scheduler clogging
     }
 }
